% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adduct-function.R
\name{chemform_adduct_check}
\alias{chemform_adduct_check}
\title{chemform_adduct_check}
\usage{
chemform_adduct_check(adduct.to.check)
}
\arguments{
\item{adduct.to.check}{}
}
\value{
df
}
\description{
check if string is a adduct (match in adduct.table)
}
