% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_envipat_lc8.R
\name{chemform_calculate_matrix}
\alias{chemform_calculate_matrix}
\title{chemical formula calculation}
\usage{
chemform_calculate_matrix(
  Formula1 = "C2H4O1S2P1",
  Formula2 = "N1H1O-1",
  sign = 1,
  Valid_formula = FALSE
)
}
\arguments{
\item{Valid_formula}{}
}
\description{
two vector of \code{chemform}, return a matrix
}
