% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_envipat_lc8.R
\name{chemform_calculate_vector}
\alias{chemform_calculate_vector}
\title{chemical formula calculation}
\usage{
chemform_calculate_vector(
  Formula1 = chem_formula_template,
  Formula2 = chem_formula_template,
  sign = sample(1, replace = T, length(Formula1))
)
}
\arguments{
\item{sign}{}
}
\description{
two vector of \code{chemform}, must be same length,return a vector
}
