% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform_calc.R
\name{chemform_sum}
\alias{chemform_sum}
\alias{chemform_multi}
\title{Chemform Calculation}
\usage{
chemform_sum(...)

chemform_multi(
  chemform = chem_formula_template,
  multi = 1,
  return = c("matrix", "chemform")
)
}
\arguments{
\item{...}{chemform list or vector}

\item{chemform}{chemform}

\item{multi}{number to multi}
}
\value{
chemform

chemform
}
\description{
Chemform Calculation

Chemform Calculation
}
\section{Functions}{
\itemize{
\item \code{chemform_sum()}: chemform_sum

\item \code{chemform_multi()}: chemform_multi

}}
\examples{
chemform_sum(chem_formula_template)
chemform_multi("C6H12O6",1)
}
