% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_envipat_lc8.R
\name{chemform_formate}
\alias{chemform_formate}
\title{chemform_formate}
\usage{
chemform_formate(chemform = "C11H22NO4")
}
\arguments{
\item{chemform}{}
}
\description{
formate a chemical formula:
\enumerate{
\item replace \code{D} with \verb{[1]H}
}
}
