% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_mz}
\alias{chemform_mz}
\title{chemform_mz}
\usage{
chemform_mz(chemform = chem_formula_template, charge = 0)
}
\arguments{
\item{chemform}{}
}
\value{
mz
}
\description{
calculate mz of a given chemform
}
