% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_parse}
\alias{chemform_parse}
\title{chemform_parse}
\usage{
chemform_parse(chemform = chem_formula_template, return = "matrix")
}
\arguments{
\item{chemform}{formula}

\item{return}{return type}
}
\value{
ele matrix of list
}
\description{
parse given chemform to a vector of elements and count
}
