% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_envipat_lc8.R
\name{isotopes_pattern_enviPat}
\alias{isotopes_pattern_enviPat}
\title{isotopes_pattern_enviPat}
\usage{
isotopes_pattern_enviPat(chemform)
}
\arguments{
\item{formula}{}
}
\description{
update of enviPat::isopattern, which do not output the formula of isotopologues
this function return isotopologues with abundance more than 0.01\%
}
