% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_NetID.R
\name{match_isoadduct_to_features}
\alias{match_isoadduct_to_features}
\title{match_adduct_to_features}
\usage{
match_isoadduct_to_features(isoadduct, xcms.xcms, ppm.thresh = 10, rt.tol = 10)
}
\arguments{
\item{ppm.thresh}{}
}
\description{
match_adduct_to_features
}
