% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adduct-function.R
\name{chemform_adduct}
\alias{chemform_adduct}
\title{chemform_adduct}
\usage{
chemform_adduct(
  chemform = chem_formula_template,
  adduct = "[M+H]+",
  value = "all"
)
}
\arguments{
\item{chemform}{chemical formula}

\item{adduct}{adduct form, such as "\link{M+H}+"}
}
\description{
get chemical formula with adduct
}
