% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_calc}
\alias{chemform_calc}
\title{chemform_calc}
\usage{
chemform_calc(
  chemform1 = chem_formula_template,
  chemform2 = rev(chem_formula_template),
  calc = "+"
)
}
\arguments{
\item{chemform1}{vector of chemform, length as m}

\item{chemform2}{vector of chemform, length as n}

\item{calc}{}
}
\value{
chemform vector or matrix
}
\description{
chemform calculator, calc "+" or "-" for vector operation, which return a vetor with length of m.
calc ".+" or ".-" for matrix operation, which return a matrix m x n
}
