% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_formate}
\alias{chemform_formate}
\title{chemform_formate}
\usage{
chemform_formate(chemform = chem_formula_template, return = "chemform")
}
\arguments{
\item{chemform}{}
}
\description{
formate a chemical formula
}
