% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_formate}
\alias{chemform_formate}
\title{chemform_formate}
\usage{
chemform_formate(chemform = chem_formula_template)
}
\arguments{
\item{chemform}{}
}
\description{
formate a chemical formula:
\enumerate{
\item add 1 after single element
\item replace \code{D} with \verb{[1]H}
}
}
