% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_get_ele}
\alias{chemform_get_ele}
\title{chemform_get_ele}
\usage{
chemform_get_ele(chemform)
}
\arguments{
\item{chemform}{}
}
\description{
char considered as a element should follow one of:
\enumerate{
\item uppercase with uppercase or number after it
\item continuous uppercase and lowercase
\item continuous [n] and uppercase with uppercase or number after it
\item continuous [n] and uppercase and lowercase
}
}
\examples{
chemform_get_ele("CCC[13]C")
}
