% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_envipat_lc8.R
\name{chemform_isotopes_pattern_enviPat}
\alias{chemform_isotopes_pattern_enviPat}
\title{chemform_isotopes_pattern_enviPat}
\usage{
chemform_isotopes_pattern_enviPat(chemform, thresh = 0.1)
}
\arguments{
\item{chemform}{chemical formula}
}
\value{
isotopes pattern
}
\description{
update of enviPat::isopattern, which do not output the formula of isotopologues
this function return isotopologues with abundance more than 0.01\%
}
