% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_envipat_lc8.R
\name{chemform_mz_lc8}
\alias{chemform_mz_lc8}
\title{chemform_mz_lc8}
\usage{
chemform_mz_lc8(chemform = "C2H4O1S2P1")
}
\arguments{
\item{chemform, }{such as \code{"C2H4"}}
}
\description{
calculate mz of a given chemical formula
}
