% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chemform-function.R
\name{chemform_parse}
\alias{chemform_parse}
\title{chemform_parse}
\usage{
chemform_parse(chemform = chem_formula_template, return = "list")
}
\arguments{
\item{return}{}
}
\description{
parse given chemform to a vector of elements and count
}
