% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_NetID.R
\name{expand_isoadduct_from_formula}
\alias{expand_isoadduct_from_formula}
\title{expand_isoadduct_from_formula}
\usage{
expand_isoadduct_from_formula(chem_formula, ion_mode = "positive")
}
\arguments{
\item{chem_formula}{a chemical formula, such as "C4H4Cl1N3O1".
some template see "MSCC::chem_formula_template"}

\item{ion_mode}{}
}
\description{
Using formula, exact.mass and ion_mode, expand adduct by enviPat::adduct
}
